requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.4.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
better-profanity==0.5.0
bleach==3.1.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==0.4.0
cmake==3.18.4.post1
colorama==0.4.3
coolname==1.1.0
cvxopt==1.2.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dict2xml==1.7.0
en-core-web-sm==2.0.0
entrypoints==0.3
ExifRead==2.1.2
flake8==3.8.3
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.29.0
h5py==2.10.0
html5lib==1.1
idna==2.10
imageio==2.9.0
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.6.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.5
nltk==3.4.5
nmslib==2.0.5
notebook==6.0.3
numpngw==0.0.8
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==3.4.10.35
opt-einsum==3.2.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.4.1
profanity-check==1.0.3
progressbar2==3.51.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycodestyle==2.6.0
pydicom==2.1.2
pydot==1.4.1
pydot-ng==2.0.0
pydotplus==2.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyproj==3.0.0.post1
pyrsistent==0.16.0
pytest==3.0.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
runx==0.0.10
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.2.0
tensorflow-addons==0.10.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.8.25
toml==0.10.0
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.28.1
traitlets==4.3.3
typed-ast==1.4.1
typeguard==2.11.1
urllib3==1.25.10
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xclib==0.96
xlrd==2.0.1
xmltodict==0.10.2
zipp==3.1.0
